requirements.txt
# 简历解析助手 - 核心依赖
# Python 3.9+

# PDF 解析
PyMuPDF>=1.23.0

# Word 文档解析  
python-docx>=1.1.0

# AI 大模型调用
litellm>=1.0.0

# Excel 处理
openpyxl>=3.1.0

# 环境变量
python-dotenv>=1.0.0

# Web 界面
streamlit>=1.28.0
